requirements.txt
python-telegram-bot==8.1.1
pymongo
bson
datetime
python-dotenv